% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_combined_matrix.R
\name{calculate_combined_matrix}
\alias{calculate_combined_matrix}
\title{Calculates both the Liklihood and Uncertainty (Missing Data) Matrix Between Datasets}
\usage{
calculate_combined_matrix(
  buccal_swabs,
  human_profiles,
  result_type,
  tolerance = 0
)
}
\arguments{
\item{buccal_swabs}{A data frame containing human genetic data}

\item{human_profiles}{A data frame containing human blood extracted from mosquitoes}

\item{result_type}{Denotes whether you want the results to be the genetic distance (1), or percentage match (2)}

\item{tolerance}{Denotes an acceptable tolerance range for values}
}
\value{
A matrix representing the likelihood of genetic marker matching between the samples in each mosquito and human.
}
\description{
This function calculates both the likelihood matrix, comparing genetic markers
between blood extracted from mosquitoes and human samples, and uncertainty matrix, evaluating the presence of missing (NA) values in genetic markers
between two datasets. The function returns a matrix where each cell has two values, the first showing liklihood, and second uncertainty in parenthesis
}
\examples{
# Example usage:
data("buccal_swabs")
data("human_profiles")
calculate_combined_matrix(buccal_swabs, human_profiles,1,0)
}
