% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_uncertainty_matrix.R
\name{calculate_uncertainty_matrix}
\alias{calculate_uncertainty_matrix}
\title{Calculate Uncertainty (Missing Data) Matrix Between Datasets}
\usage{
calculate_uncertainty_matrix(buccal_swabs, human_profiles)
}
\arguments{
\item{buccal_swabs}{A data frame containing human genetic data from buccal swabs.}

\item{human_profiles}{A data frame containing genetic data extracted from mosquitoes, presumably after biting humans.}
}
\value{
A matrix where rows correspond to human samples and columns correspond to mosquito samples.
        Each cell in the matrix shows the percentage of genetic markers that have missing data in either or both
        of the paired samples.
}
\description{
This function calculates an uncertainty matrix by evaluating the presence of missing (NA) values in genetic markers
between two datasets: one representing human genetic data from buccal swabs and the other representing genetic data
extracted from mosquitoes. The function returns a matrix where each element represents the percentage of missing data
for the corresponding pair of human and mosquito samples.
}
\examples{
# Example usage:
data("buccal_swabs")  # Load buccal swab data
data("human_profiles")  # Load mosquito extracted genetic data
missing_data_matrix <- calculate_uncertainty_matrix(buccal_swabs, human_profiles)

}
