% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_missing_data.R
\name{remove_missing_data}
\alias{remove_missing_data}
\title{Remove Missing Data}
\usage{
remove_missing_data(data, type, threshold)
}
\arguments{
\item{data}{The input dataset.}

\item{type}{The type of missing data removal. Choose between "column" or "row".}

\item{threshold}{The threshold percentage of missing data above which columns or rows will be removed.}
}
\value{
The modified dataset with missing data removed.
}
\description{
Remove columns or rows with missing data exceeding a specified threshold percentage.
}
\examples{
# Remove columns with missing data exceeding 20\%
new_data <- remove_missing_data(data, "column", 0.2)

# Remove rows with missing data exceeding 30\%
new_data <- remove_missing_data(data, "row", 0.3)

}
