% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_columns.R
\name{sort_columns}
\alias{sort_columns}
\title{Sort Columns in data set}
\usage{
sort_columns(unordered_data)
}
\arguments{
\item{unordered_data}{Data that needs to be ordered}
}
\value{
Data ordered in ASC order
}
\description{
Sorts all columns (except the first) in each sheet of an Excel file in ascending order and writes the sorted data to a new Excel file.
}
\examples{
# Sort columns in a given data set
sort_columns(dataset)

}
