% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_excel_column.R
\name{sort_excel_column}
\alias{sort_excel_column}
\title{Sort Columns in Excel Sheets}
\usage{
sort_excel_column(file_path, output_file)
}
\arguments{
\item{file_path}{Path to the input Excel file.}

\item{output_file}{Path to the output Excel file where the sorted data will be saved.}
}
\value{
None. The sorted data is written to the output Excel file.
}
\description{
Sorts all columns (except the first) in each sheet of an Excel file in ascending order and writes the sorted data to a new Excel file.
}
\examples{
# Sort columns in input.xlsx and save the sorted data to output.xlsx
sort_columns("input.xlsx", "output.xlsx")

}
