% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_likelihood_matrix.R
\name{calculate_likelihood_matrix}
\alias{calculate_likelihood_matrix}
\title{Calculate likelihood matrix based on genetic marker matching}
\usage{
calculate_likelihood_matrix(volunteer_samples, human_samples_in_mosquito)
}
\arguments{
\item{volunteer_samples}{A data frame containing human genetic data}

\item{human_profiles}{A data frame containing human blood extrated from mosquitoes}
}
\value{
A matrix representing the likelihood of genetic marker matching between the samples in each mosquito and human.
}
\description{
This function calculates the likelihood matrix by comparing genetic markers
between blood extracted from mosquitoes and human samples.
}
\examples{
# Example usage:
data("buccal_swabs")
data("human_profiles")
calculate_new_likelihood_matrix(buccal_swabs, human_profiles)
}
